# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * WebIP API specification - IP lookup and website inspection
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface IpInfo {
  ip: string;
  /** @nullable */
  ipVersion?: string | null;
  /** @nullable */
  hostname?: string | null;
  /** @nullable */
  reverseDns?: string | null;
  /** @nullable */
  isp?: string | null;
  /** @nullable */
  asn?: string | null;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  continent?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  countryCode?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  googleMapsUrl?: string | null;
  /** @nullable */
  currencyName?: string | null;
  /** @nullable */
  currencyCode?: string | null;
  /** @nullable */
  languages?: string | null;
  /** @nullable */
  network?: string | null;
  /** @nullable */
  connectionType?: string | null;
  /** @nullable */
  hostingProvider?: boolean | null;
  /** @nullable */
  isProxy?: boolean | null;
  /** @nullable */
  whois?: string | null;
}

export interface KeyValue {
  key: string;
  value: string;
}

export interface DnsRecords {
  a?: string[];
  aaaa?: string[];
  mx?: string[];
  ns?: string[];
  txt?: string[];
  cname?: string[];
}

export interface SslInfo {
  valid?: boolean;
  /** @nullable */
  issuer?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  validFrom?: string | null;
  /** @nullable */
  validTo?: string | null;
  /** @nullable */
  daysRemaining?: number | null;
  /** @nullable */
  protocol?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface HeadingGroup {
  h1?: string[];
  h2?: string[];
  h3?: string[];
  h4?: string[];
  h5?: string[];
  h6?: string[];
}

export type SeoFindingSeverity = typeof SeoFindingSeverity[keyof typeof SeoFindingSeverity];


export const SeoFindingSeverity = {
  good: 'good',
  warning: 'warning',
  error: 'error',
} as const;

export interface SeoFinding {
  severity: SeoFindingSeverity;
  message: string;
}

export interface WebsiteInspection {
  requestedUrl: string;
  finalUrl: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  canonicalUrl?: string | null;
  /** @nullable */
  metaRobots?: string | null;
  openGraph?: KeyValue[];
  twitterCard?: KeyValue[];
  /** @nullable */
  charset?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  favicon?: string | null;
  httpStatus: number;
  responseHeaders?: KeyValue[];
  redirectChain?: string[];
  sslInfo?: SslInfo;
  dns?: DnsRecords;
  /** @nullable */
  registrarInfo?: string | null;
  /** @nullable */
  robotsTxt?: string | null;
  /** @nullable */
  sitemapXml?: string | null;
  internalLinks?: string[];
  externalLinks?: string[];
  cssFiles?: string[];
  jsFiles?: string[];
  images?: string[];
  fonts?: string[];
  formsCount?: number;
  buttonsCount?: number;
  tablesCount?: number;
  headings?: HeadingGroup;
  structuredData?: string[];
  jsonLd?: string[];
  technologies?: string[];
  analyticsDetected?: string[];
  seoAudit?: SeoFinding[];
  accessibilitySuggestions?: SeoFinding[];
  performanceSuggestions?: SeoFinding[];
  htmlSource?: string;
}

export type LookupIpParams = {
ip?: string;
};

export type InspectWebsiteParams = {
url: string;
};