# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * WebIP API specification - IP lookup and website inspection
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
}