# lib/api-zod/src/generated/types/headingGroup.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * WebIP API specification - IP lookup and website inspection
 * OpenAPI spec version: 0.1.0
 */

export interface HeadingGroup {
  h1?: string[];
  h2?: string[];
  h3?: string[];
  h4?: string[];
  h5?: string[];
  h6?: string[];
}