# lib/api-zod/src/generated/types/ipInfo.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * WebIP API specification - IP lookup and website inspection
 * OpenAPI spec version: 0.1.0
 */

export interface IpInfo {
  ip: string;
  /** @nullable */
  ipVersion?: string | null;
  /** @nullable */
  hostname?: string | null;
  /** @nullable */
  reverseDns?: string | null;
  /** @nullable */
  isp?: string | null;
  /** @nullable */
  asn?: string | null;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  continent?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  countryCode?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  googleMapsUrl?: string | null;
  /** @nullable */
  currencyName?: string | null;
  /** @nullable */
  currencyCode?: string | null;
  /** @nullable */
  languages?: string | null;
  /** @nullable */
  network?: string | null;
  /** @nullable */
  connectionType?: string | null;
  /** @nullable */
  hostingProvider?: boolean | null;
  /** @nullable */
  isProxy?: boolean | null;
  /** @nullable */
  whois?: string | null;
}