# lib/api-zod/src/generated/types/seoFinding.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * WebIP API specification - IP lookup and website inspection
 * OpenAPI spec version: 0.1.0
 */
import type { SeoFindingSeverity } from './seoFindingSeverity';

export interface SeoFinding {
  severity: SeoFindingSeverity;
  message: string;
}