# lib/api-zod/src/generated/types/sslInfo.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * WebIP API specification - IP lookup and website inspection
 * OpenAPI spec version: 0.1.0
 */

export interface SslInfo {
  valid?: boolean;
  /** @nullable */
  issuer?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  validFrom?: string | null;
  /** @nullable */
  validTo?: string | null;
  /** @nullable */
  daysRemaining?: number | null;
  /** @nullable */
  protocol?: string | null;
  /** @nullable */
  error?: string | null;
}