# lib/api-zod/src/generated/types/websiteInspection.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * WebIP API specification - IP lookup and website inspection
 * OpenAPI spec version: 0.1.0
 */
import type { DnsRecords } from './dnsRecords';
import type { HeadingGroup } from './headingGroup';
import type { KeyValue } from './keyValue';
import type { SeoFinding } from './seoFinding';
import type { SslInfo } from './sslInfo';

export interface WebsiteInspection {
  requestedUrl: string;
  finalUrl: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  canonicalUrl?: string | null;
  /** @nullable */
  metaRobots?: string | null;
  openGraph?: KeyValue[];
  twitterCard?: KeyValue[];
  /** @nullable */
  charset?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  favicon?: string | null;
  httpStatus: number;
  responseHeaders?: KeyValue[];
  redirectChain?: string[];
  sslInfo?: SslInfo;
  dns?: DnsRecords;
  /** @nullable */
  registrarInfo?: string | null;
  /** @nullable */
  robotsTxt?: string | null;
  /** @nullable */
  sitemapXml?: string | null;
  internalLinks?: string[];
  externalLinks?: string[];
  cssFiles?: string[];
  jsFiles?: string[];
  images?: string[];
  fonts?: string[];
  formsCount?: number;
  buttonsCount?: number;
  tablesCount?: number;
  headings?: HeadingGroup;
  structuredData?: string[];
  jsonLd?: string[];
  technologies?: string[];
  analyticsDetected?: string[];
  seoAudit?: SeoFinding[];
  accessibilitySuggestions?: SeoFinding[];
  performanceSuggestions?: SeoFinding[];
  htmlSource?: string;
}